10/10 new

+++ src/main/java/com/example/hiepphat/repositories/LikeBlogRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface LikeBlogRepository extends JpaRepository<LikeBlog,Long> {
-    LikeBlog findByUser_UserIDAndBlog_BlogID(int user_id,long blog_id);
+public interface LikeBlogRepository extends JpaRepository<LikeBlog,Integer> {
+    LikeBlog findByUser_UserIDAndBlog_BlogID(int user_id,int blog_id);
     List<LikeBlog> findByBlog_BlogID(int id);
 }
