+++ src/main/java/com/example/hiepphat/service/LikeBlogService.java
@@ -3,6 +3,6 @@ package com.example.hiepphat.service;
 import com.example.hiepphat.Entity.LikeBlog;
 
 public interface LikeBlogService {
-    LikeBlog findByUser_UserIDAndBlog_BlogID(int user_id, long blog_id);
+    LikeBlog findByUser_UserIDAndBlog_BlogID(int user_id, int blog_id);
     void deleteLike(int id);
 }
